+++ src/components/dashboard/cars/components/DeleteConfirmationDialog.tsx
@@ -1,0 +1,51 @@
+// src/components/dashboard/cars/components/DeleteConfirmationDialog.tsx
+"use client";
+
+import React from "react";
+import { useTranslations } from "next-intl";
+import { Button } from "@/components/ui/button";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+
+interface DeleteConfirmationDialogProps {
+  isOpen: boolean;
+  onClose: () => void;
+  onConfirm: () => void;
+}
+
+const DeleteConfirmationDialog: React.FC<DeleteConfirmationDialogProps> = ({
+  isOpen,
+  onClose,
+  onConfirm,
+}) => {
+  const t = useTranslations("dashboard");
+
+  return (
+    <Dialog open={isOpen} onOpenChange={onClose}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>{t("cars.deleteConfirmation.title")}</DialogTitle>
+          <DialogDescription>
+            {t("cars.deleteConfirmation.description")}
+          </DialogDescription>
+        </DialogHeader>
+        <DialogFooter>
+          <Button variant="outline" onClick={onClose}>
+            {t("cars.deleteConfirmation.cancel")}
+          </Button>
+          <Button variant="destructive" onClick={onConfirm}>
+            {t("cars.deleteConfirmation.confirm")}
+          </Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+};
+
+export default DeleteConfirmationDialog;
